refactor: Timer 컴포넌트를 shared로 이동
