+++ frontend/src/components/@shared/Title/index.tsx
@@ -3,8 +3,9 @@ import { PropsWithChildren } from 'react';
 import { MdArrowBackIosNew } from 'react-icons/md';
 import styled, { css } from 'styled-components';
 
+import { TitleProps, TitleWrapperProps } from 'components/@shared/Title/type';
+
 import { FlexBox, Text } from 'components';
-import { TitleProps, TitleWrapperProps } from 'components/Title/type';
 
 import { Z_INDEX } from 'constants/css';
 import { TITLE_HEIGHT } from 'constants/style';
