+++ frontend/src/components/@shared/Title/type.ts
