+++ frontend/src/components/index.ts
@@ -20,6 +20,7 @@ export * from 'components/@shared/UnderLineText';
 export * from 'components/@shared/Tabs';
 export * from 'components/@shared/SnackBar';
 export * from 'components/@shared/SROnly';
+export * from 'components/@shared/Title';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
 export * from 'components/Header';
@@ -34,7 +35,6 @@ export * from 'components/Profile';
 export * from 'components/EmptyContent';
 export * from 'components/PrivateOutlet';
 export * from 'components/LandingNavigation';
-export * from 'components/Title';
 export * from 'components/Challenger';
 export * from 'components/ChallengerList';
 export * from 'components/ChallengeDetail';
